// src/app/models/course.model.ts
export interface Course {
  id: string;
  name: string;
  remainingSpots?: string;
  department?: string;
  url?: string;
  educationalComponent?: string;
  teacher?: string;
  requirements?: string;
  whatWillBeStudied? : string;
  whyIsThisInteresting?: string;
  learningOutcomes?: string;
  howToUseKnowledge?: string;
}